#85 feat: sideBar 수정
- 타입 및  컴포넌트 분리
- 구조 수정
- dummy 제거
