+++ frontend/src/constant/CheckBox.ts
