+++ frontend/src/constant/ElementType.ts
@@ -1,0 +1,8 @@
+enum ElementType {
+  AddButton,
+  Assignee,
+  Label,
+  milestone,
+}
+
+export default ElementType;
