+++ frontend/src/pages/Components.tsx
@@ -1,8 +1,8 @@
 import { styled } from 'styled-components';
-import Button from '../components/common/button/BaseButton';
-import ButtonLarge from '../components/common/button/ButtonLarge';
-import ButtonSmall from '../components/common/button/ButtonSmall';
-import ColorCodeInput from '../components/common/ColorCodeInput';
+// import Button from '../components/common/button/BaseButton';
+// import ButtonLarge from '../components/common/button/ButtonLarge';
+// import ButtonSmall from '../components/common/button/ButtonSmall';
+// import ColorCodeInput from '../components/common/ColorCodeInput';
 // import TabButton from '../components/common/TabButton';
 // import InformationTag from '../components/common/InformationTag';
 // import TextInput from '../components/common/TextInput';
@@ -15,6 +15,10 @@ import ColorCodeInput from '../components/common/ColorCodeInput';
 // import Comment from '../components/common/CommentElements';
 import { AppContext } from '../main';
 import { useContext } from 'react';
+import SideBar from '../components/common/sideBar/SideBar';
+import ElementType from '../constant/ElementType';
+
+const { AddButton, Assignee, Label } = ElementType;
 
 export default function Components() {
   const appContext = useContext(AppContext);
@@ -22,7 +26,7 @@ export default function Components() {
   return (
     <StyledComponents>
       <Wrapper>
-        <ButtonLarge type="submit" iconName="plus">
+        {/* <ButtonLarge type="submit" iconName="plus">
           BUTTON
         </ButtonLarge>
         <Button type="submit" outline iconName="plus">
@@ -41,7 +45,7 @@ export default function Components() {
           S BUTTON
         </ButtonSmall>
         <ColorCodeInput label="배경색" />
-        {/* <TabButton />
+        <TabButton />
         <InformationTag size="medium" iconName="alertCircle">
           Label
         </InformationTag>
@@ -76,8 +80,24 @@ export default function Components() {
           timeStamp="5분 전"
           comment="오늘 점심 뭐 먹죠?"
         />
-        <SideBar />
         <FilterBar /> */}
+        <SideBar
+          sideBarItems={[
+            [{ type: AddButton, text: '담당자' }, []],
+            [
+              { type: AddButton, text: '레이블' },
+              [{ type: Assignee, text: 'label2', checked: false }],
+            ],
+            [
+              { type: AddButton, text: '마일스톤' },
+              [{ type: Label, labelProps: {
+                textColor: 'AAA333',
+                backgroundColor: 'FFF3FF',
+                name: 'refactor',
+              } }],
+            ],
+          ]}
+        />
       </Wrapper>
     </StyledComponents>
   );
