+++ frontend/src/types/LabelProps.ts
@@ -1,0 +1,7 @@
+type LabelProps = {
+  textColor: string;
+  backgroundColor: string;
+  name: string;
+};
+
+export default LabelProps
